+++ install-assets/badger_initial_security_pgsql.sql
@@ -3,6 +3,7 @@ DROP SEQUENCE IF EXISTS element_id_seq;
 CREATE SEQUENCE element_id_seq;
 CREATE TABLE co_security_nodes (
   id BIGINT NOT NULL DEFAULT nextval('element_id_seq'),
+  api_key VARCHAR(255) DEFAULT NULL,
   login_id VARCHAR(255) DEFAULT NULL,
   access_class VARCHAR(255) NOT NULL,
   last_access TIMESTAMP NOT NULL,
@@ -13,5 +14,5 @@ CREATE TABLE co_security_nodes (
   ids VARCHAR(4095) DEFAULT NULL
 );
 
-INSERT INTO co_security_nodes (login_id, access_class, last_access, read_security_id, write_security_id, object_name, access_class_context, ids) VALUES
-(NULL, 'CO_Security_Node', '1970-01-01 00:00:00', -1, -1, NULL, NULL, NULL);
+INSERT INTO co_security_nodes (api_key, login_id, access_class, last_access, read_security_id, write_security_id, object_name, access_class_context, ids) VALUES
+(NULL, NULL, 'CO_Security_Node', '1970-01-01 00:00:00', -1, -1, NULL, NULL, NULL);