+++ install-assets/s_config.class.php
@@ -26,7 +26,20 @@ class CO_Config {
     /// These affect the overall "God Mode" login.
     static private $_god_mode_id            = 2;                        ///< God Login Security DB ID. This is private, so it can't be programmatically changed.
     static private $_god_mode_password      = '<GOD MODE PASSWORD>';    ///< Plaintext password for the God Mode ID login. This overrides anything in the ID row.
-    
+    static private $_log_handler_function = NULL;                       /**<    This is a special callback for logging REST calls (BASALT). For most functions in the global scope, this will simply be the function name,
+                                                                                or as an array (with element 0 being the object, itself, and element 1 being the name of the function).
+                                                                                If this will be an object method, then it should be an array, with element 0 as the object, and element 1 a string, containing the function name.
+                                                                                The function signature will be:
+                                                                                    function log_callback ( $in_andisol_instance,   ///< REQUIRED: The ANDISOL instance at the time of the call.
+                                                                                                            $in_server_vars         ///< REQUIRED: The $_SERVER array, at the time of the call.
+                                                                                                            );
+                                                                                There is no function return.
+                                                                                The function will take care of logging the REST connection in whatever fashion the user desires.
+                                                                                This will assume a successful ANDISOL instantiation, and is not designed to replace the traditional server logs.
+                                                                                It should be noted that there may be legal, ethical and resource ramifications for logging.
+                                                                                It is up to the implementor to ensure compliance with all constraints.
+                                                                        */
+
     /// These are the basic operational settings.
     static $lang                            = 'en';                     ///< The default language for the server.
     static $min_pw_len                      = 8;                        ///< The minimum password length.
@@ -50,12 +63,10 @@ class CO_Config {
     static $sec_db_type                     = '<mysql or pgsql>';
     static $sec_db_login                    = '<SECURITY DB LOGIN>';
     static $sec_db_password                 = '<SECURITY DB PASSWORD>';
-
-    /**
-    This is the Google API key. It's required for CHAMELEON to do address lookups and other geocoding tasks.
-    CHAMELEON requires this to have at least the Google Geocoding API enabled.
-    */
-    static $google_api_key                  = '<YOUR API KEY>';
+    
+    static $google_api_key                  = '<YOUR API KEY>';         /**<    This is the Google API key. It's required for CHAMELEON to do address lookups and other geocoding tasks.
+                                                                                CHAMELEON requires this to have at least the Google Geocoding API enabled.
+                                                                        */
     
     /***********************/
     /**